+++ packages/core/src/bricks/__init__.py
@@ -2,6 +2,6 @@
 
 from bricks.api import Bricks
 
-__version__ = "0.4.36"
+__version__ = "0.4.37"
 
 __all__ = ["Bricks", "__version__"]